+++ src/app/features/home/capture-tab/capture-item/capture-item.component.ts
@@ -71,10 +71,7 @@ export class CaptureItemComponent {
     )
   );
 
-  readonly hasCaption$ = this.proof$.pipe(
-    switchMap(proof => this.diaBackendAssetRepository.fetchByProof$(proof)),
-    map(asset => asset.caption !== '')
-  );
+  readonly hasCaption$ = this.proof$.pipe(map(proof => proof.caption !== ''));
 
   readonly isVideo$ = this.proof$.pipe(
     concatMap(proof => proof.getFirstAssetMeta()),