+++ src/app/shared/dia-backend/asset/downloading/dia-backend-downloading.service.ts
@@ -80,6 +80,8 @@ export class DiaBackendAssetDownloadingService {
       },
     });
     proof.diaBackendAssetId = diaBackendAsset.id;
+    proof.caption = diaBackendAsset.caption;
+    proof.uploadedAt = diaBackendAsset.uploaded_at;
     if (diaBackendAsset.signed_metadata) proof.setSignatureVersion();
     return this.proofRepository.add(proof, OnConflictStrategy.REPLACE);
   }